+++ Plugins/FIFA23Plugin/SDKGen/FieldInfo.cs
@@ -7,8 +7,7 @@ namespace SdkGenerator.Fifa23
 {
 	public class FieldInfo : IFieldInfo
 	{
-		//public uint nameHash;
-		public int nameHash { get; set; }
+		public uint nameHash { get; set; }
 
 		public static Random RandomEmpty = new Random();
 
@@ -41,8 +40,7 @@ namespace SdkGenerator.Fifa23
 			{
 				name = parentTypeInfo.name + "_UnkField_" + RandomEmpty.Next().ToString();
 			}
-			//nameHash = reader.ReadUInt();
-			nameHash = (int)reader.ReadUInt();
+			nameHash = reader.ReadUInt();
 			flags = reader.ReadUShort();
 			offset = reader.ReadUShort();
 			typeOffset = reader.ReadLong();
